typing: add typehints to mercurial/exthelper.py
This helps a little (there were some things flagged in largefiles when I had
something hinted wrong), but probably not a lot because pytype struggles with
method decorators, and we don't use this class a lot.  But it does show in hover
tips in PyCharm for the decorators, so it's somewhat helpful.

--HG--
extra : rebase_source : 26b9deb4e3491ed03fe8fc8ec28a92a1e4f23e84
extra : topic : typing-exthelper