typing: disable a missing attribute error around `RepoT` in remotefilelog
This will be caught by pytype once `RepoT` is aliased to the repository protocol
class instead of Any.  We don't currently have an extension specific subclass
for here, and it looks like this attribute is just slapped onto the normal repo
instance (like some attrs related to sub-repositories are added to the plain
localrepo class).  Given that, we just disable the diagnostic to get it out of
the way.

--HG--
extra : amend_source : ec044d815179f64c83401366cb5939267a204626
extra : topic : typing-repo-alias