typing: add signatures for extension loading hooks
Some of the `serve` related code touched in c011800746bb deals with methods
related to `reposetup()`, and some have open coded typing.  Let's add something
easy to reference instead.

I'm not sure that this is the right module (as opposed to `_basetypes.py`), but
we already know we'll have to import the module with the `ui` protocol class,
and `_basetypes.py` isn't supposed to import these other interface modules.  So
start it here, and we'll deal with issues if/when they arise.

--HG--
extra : rebase_source : daaf5276c11c4dadc0a656a8bdfc7bd3ca1515ac
extra : topic : typing-repo-alias