typing: drop a `disable=pyi-error` block in `mercurial.hgweb`
Maybe this was because of a buggy earlier version, or maybe it was because of
all of the cycles until recently.  Whatever the case, it doesn't fail now.  This
also doesn't alter any of the generated pyi files.

--HG--
extra : rebase_source : dd0c81f291215436c1f251546679068a13b8030e
extra : topic : typing-repo-alias