rust-delta: provides line_diff in Rust
Having the logic in rust will allow for more flexible usage and further
performance optimization.

I am using the "imara-diff" crate before it claimed to be fast and its API
provided something close to what we needed. There as a few things that seems
suboptimal in its implementation (e.g. the hashing is probably more complicated
than we need), but it already provides better performance than the C
implementation of bdiff used so far and the API is simple enough.

In its current form, the Rust implementation is a bit slower than calling the
existing C, because we don't do the prefix/suffix trimming in Rust yet.  This
will be fixed in the next two changesets.

There are a few TODO left for when we want to dig deeper. However in its current
form, this will behave well enough for other part of the code to now dominates
the problem space.  (like building the full text to build)

--HG--
extra : rebase_source : cde36ed0469e329cfabdf1881a99fb14961179d3
extra : topic : delta-rust