typing: cleanup type checking issues in `mercurial/dispatch.py`
Now that we dropped support for Python 3.6, `initstdio()` can be refactored.
But that's not relevant to the typing issue here, where for some reason the
constructor wants a bool, but the attribute on the object is an int.

--HG--
extra : rebase_source : ade630b37df3f801b72fc764724b645ee3b6a90e
extra : topic : typing-repo-alias