fileindex: implement garbage collection
This makes the file index clean up old files instead of leaving them forever.

When the file index rewrites a data file instead of appending, it chooses a
filename with a new ID. This happens for the tree file when vacuuming, and for
all data files when narrowing or stripping. Previously the old files were left
on disk forever. Now, we add their paths to a list in the docket. There are two
ways they can then be deleted:

* The hg debug::file-index --gc command deletes them all immediately.

* Otherwise, every transaction scans the list and deletes files if a sufficient
  number of transactions and duration of time have passed.

This ensures that we can roll back the transaction, and that other processes can
continue reading the old file (particularly if the repo is on NFS, which can
produce a stale file handle error if an open file is unlinked).

The time duration is configurable with storage.fileindex.gc-retention-seconds,
and defaults to one day (86400).

--HG--
extra : rebase_source : 4a490b6702741ff528036e1a9a13a175cc24baac
extra : topic : file-index.trash