typing: preserve the method signature across `util.checksignature()`
This may not make much difference because it is being used inside the extension
wrapping code, but it won't hurt to have it.  Ultimately, the changes here
didn't affect the generated *.pyi file (aside from this method itself), so it's
not possible to know if the issue mentioned in f093b740ce3f would be in play
here too.  Play it safe and do it the same way, instead of using `ParamSpec`.

--HG--
extra : rebase_source : e9e50427a1004f0fa930807bab41970ce6b2fb4a
extra : topic : typing-exthelper