lfs: correct the name of the `changegroup.allsupportedversions()` arg
No idea if this was always wrong, or if the wrapped method got refactored, but
it takes a `RepoT`, not a `UiT`.  This was found manually when trying to adapt
the decorator to preserve the original signature, and possibly check the
signature of the wrapped function.  There was no bug here, since the argument
was passed to the wrapped method unchanged, but it was misleading.

While here, add the appropriate type hints.

--HG--
extra : rebase_source : b751e3f8d6e54ec43ba3c2ddb08d6f329d24deca
extra : topic : typing-exthelper