file-index: use ui.formatter debug::file-index --docket
This addresses the comment from a while ago on !1559:
https://foss.heptapod.net/mercurial/mercurial-devel/-/merge_requests/1559#note_484649

It makes no difference to the default output, but it means built-in templates
like -Tjson work now. Note that -Tjson is hardcoded to always emit an array, so
in this case it's always a singleton array.

The motivation for this is that I'm planning to add a docket field which is a
list of dicts (garbage entries), and formatting that correctly seems easier when
using the ui.formatter API.

--HG--
extra : rebase_source : 28e2ee3b65acb7408e277ff3622ebc3f0dad7613
extra : topic : file-index.trash