narrow: introduce a new way of representing the narrowspec
The current way we process the narrowspec does not allow for proper
composition of includes and excludes and prevents us from expressing very
useful patterns.

In this change, we start using a tree which minimally represents the narrowspec
while not changing anything for now. It will enable us to create fingerprints
for narrowspecs by being a form of normalization, and in turn enable the
mechanism of shapes, to be introduced in the coming patches.

--HG--
extra : rebase_source : 8533435cec915e61c67e35dc2223dfc730b5599f
extra : topic : narrow-shape.fingerprint