typing: add typehints to the lfs wrapping of `engine.finishdatamigration()`
This is the wrapper around the method in the previous commit.

I don't see any (working) way to enforce that the `orig` callable signature
matches the subsequent arguments, and that that signature also matches the
wrapped function.  Maybe there are some inspection methods that can be called to
compare signatures when the wrapping is done, and only done during test runs, in
order to catch all signature changes in core.  But for now, we can document the
current expectations.

--HG--
extra : rebase_source : db6f33579a0e4c1ff57fc3a6efafc6709c024dd4
extra : topic : typehint-upgrade-utils