delta-fold: consume the source in the `combine` method
The method take &mut value because it modify the DeltaPiece in place. So after
the apply the Delta are not expected to be usable. We clarify this fact by
consuming the input in that function.

--HG--
extra : rebase_source : 99ad7367103d56c52df6d4a7edb4707745815b05
extra : topic : revdiff.delta-consume