typing: preserve the signature of `@util.lrucachefunc` methods more accurately
The typehint bug in the transaction journaling that was fixed in f69ec6961c7e
was undetected with `ParamSpec` typing, but was found like this.  So while it
should be the same, it's obviously not to pytype.

--HG--
extra : rebase_source : e017ad226234aaa8e559c7f7d17c067c8697c288
extra : topic : typing-callable3