revision-cache: add assertion that "node" and "rev" as always in sync
Having both `node` and `rev` is source of subtle fragility lead to not using the
cache in various case.

In practice this belt-and-suspenders approach dates back from a much older time
and is no longer relevant now.

In practice, if there was situation where they diverge, we would notice as the
revision cache is used from it's revnum value only when used as a "ladder" when
restoring a revision from its delta chain.

So we add these assert in dedicated changeset to show that all is okay because
simplifying the cache.

--HG--
extra : rebase_source : 944a919e0694e32855ce4f9f43fb41af07779e0f
extra : topic : revlog.cache