revision-cache: keep the validated status in the cache
It turns out that we were already storing potential unvalidated text in that
cache (when caching it on the fly during exchange).

We don't propane this to the Rust part yet as it currently has a more limited
usage of the cache.

--HG--
extra : rebase_source : 80b78ef03d446c5daeca1f587256231bf03918f2
extra : topic : revlog.cache