raw-data: use a `Rawdata` type to abstract the current `Arc<[u8]>` usage
This type goal is to unify the type declaration in various places that want to
express "data for a revision". This will clarify the API and make it simpler to
unify the type used accross the codebase and helps with changing the type over
time.

--HG--
extra : rebase_source : 28b2a0df68bce2dbb8518d57900e7ad755e9fb47
extra : topic : revlog.raw-data