util: unconditionally use `time.perf_counter()` as the performance timer
As far as I can tell, this is was an attempt to use a py3 construct, and fall
back to the py2 `time.{time,clock}` if it wasn't available.  The documentation
says it was introduced in Python 3.3, with no mention of lacking support on some
OSes.

--HG--
extra : rebase_source : 544fd40ee81e176450fe1056f8b102447838c425
extra : topic : typing-callable3