typing: add typehints around external commands in the subrepo module
This makes PyCharm happy, and fixes about 10 warnings in it.  I pulled it out of
the previous commit because there's some trickery worth commenting on.

The big issue is the order of the `@overload` seems to need to be with the
default arg value coming first.  Even with two overloads for exactly True and
False, PyCharm was unable to understand that when the `stream` value is not
explicitly provided, it's the equivalent of passing the literal in the default
arg.  Most of the places this is called do not explicitly turn off streaming, so
it was seeing stream object values returned when it should have been seeing
bytes.  I didn't the other way around in pytype, but pytype is happy with this
order too because it now infers a bunch of other things correctly.

Second, even though there are exactly two overloads defined, PyCharm isn't able
to figure out that only those two values are acceptable when passing along to
another method with exactly those two literals, and complains.  Therefore, the
methods that are called by another method also have an overload with a
non-literal value.

Third, PyCharm seems to think that the values returned from `popen` are
str-related, not bytes.  Perhaps it's on to something because the documentation
says:

    text – If true, decode stdin, stdout and stderr using the given encoding
           (if set) or the system default otherwise.

The `text` arg is a tri-state, and something I read made me think this would
either decode or not based on some system default.  Are we sure this is always
bytes if we don't specify `text=False`?

--HG--
extra : rebase_source : 75ef10a0da0ce1a9c03217b90aac63b42ad89060
extra : topic : typing-callable3