rev-diff: move some of the logic in an enum/struct holding some of the data
The struct is able to hold enough data to just return what we need without the scoped variable
This also split the code clarity.

This is be especially useful as we are about to add more branching for code
dedicated to the manifest.

--HG--
extra : rebase_source : 3a6a814cf31633103646c452d05f9af25868393e
extra : topic : revdiff.single-fetch