util: add a decorator for emitting deprecation warnings
This is a little clearer for existing methods, given IDE syntax highlighting,
and is simpler than rolling a trivial kwarg method forwarding method to call the
warning emitter when the method is moved entirely.  This also preserves the
signature in the latter case, on the off chance the new method changes its
signature (which isn't covered by kwarg style invocation).

There doesn't seem to be a need for the `stacklevel`, since this is either a
decorator or used as a manual wrapper.  But it can be added later if there's a
use.

--HG--
extra : rebase_source : d0f6f03ceb46ce892a1277075490a2b280b518b3
extra : topic : bundle-signature