file-index: allow plain encoding
This makes the requirement "exp-fileindex-v1" compatible with the requirement
"exp-very-fragile-and-unsafe-plain-store-encoding". With it, the file index uses
plain encoding. Without it, the file index continues to use dotencode encoding
(and continues to be incompatible with an explicit "dotencode" requirement).

Since fileindex radix storage obsoletes this encoding, the original plan was
to just hardcode it to use dotencode in the meantime, and never allow specifying
encodings via requirements. However, now that we also have plain encoding, we
want to be able to migrate plain encoded repos to the file index.

--HG--
extra : rebase_source : 8711b28383998c17483d12ad10673d27d2e1befc
extra : topic : file-index.minor2