delta-info: inherit the delta-info from the cached delta if any
Propagating delta-info from one peer to the other is useful. This currently only
apply to the usage of `revlog.clone`, but this open the way to exchanging the
information in "changegroup-v4".

This needs more configurability but this will be dealt with later.

--HG--
extra : rebase_source : 2ee68cab87d232db551287bc4c92bd112902a4af
extra : topic : delta-info-quality