rust-file-index: add unit tests
This adds tests in Rust for the file index. Although it is mostly tested in
test-file-index.t, that does not cover behavior of the FileIndex object when
changes are in memory and not yet written.

This also revealed a bug where calling `iter` between `remove` and `write` gives
the wrong results. I will fix it in a follow-up changeset.

--HG--
extra : rebase_source : 25356b9b898b1d2aa00194deb53131f6f6bd7acd
extra : topic : file-index.test