rust: bump pyo3 version to 0.27.1
We need this to Python 3.14 compatibility.

This introduce the minimal amount of changes for the code to compile and the
mercurial test to pass. This makes cargo-clippy unhappy because of all the
warning, so we drop that constraints for this changeset only.

We fix them in the next changesets to avoid the noise in this one.

--HG--
extra : rebase_source : 44708522700bad8449201a2c1619cac1e1bb0367
extra : topic : python-3.14