rebase: commit rebased changesets using a filtered repo
Here, `repo` is a repository returned by `rebaseruntime.repo` property, which
can return an unfiltered repository. The reason to ever operate on an
unfiltered repo is to avoid hash visibility issues (see rebaseruntime.__init__
method).

The issue is during the committing we would start a transaction on an
unfiltered repo too, and that might be incorrect in itself, especially since
the user didn't ask for an unfiltered repo explicitly. Transaction on an
unfiltered repo also doesn't work correctly with topic extension, because topic
doesn't handle unfiltered repos (or rather it assumes that they don't have any
topics).

On the other hand, here we should already have all the data required for
committing and not have any visibility issues, so providing the original,
filtered, repo via `self._repo` to the `rebase.commit*` functions to start a
transaction and operate on works correctly. At least I think so, based on the
tests.

For the record, if we make `rebaseruntime.repo` return the original filtered
repo in all cases, there are failures, so apparently there are tests that check
hash visibility during rebase.

--HG--
extra : rebase_source : 55750d1342da323f1e2be8c82740960622e90237
extra : topic : rebase-filtered