ci: work around setuptool-scm shortcoming with pure install
We have a chicken and egg issue with Python3.14 given that we never built wheel
for it. The CI images currently use a "pure" install that does not support zstd
and struggle with the repository in the CI. This only affect test-install.t and we
work around this temporarily by skipping the setuptool-scm part of the install.

That seems like good enough for now. We can drop it once a wheel exist and the
CI image can be rebuilt.

Strictly speaking, we could change the CI image to install Mercurial from source
just for 3.14, but that does seems to be worth the trouble for now.

--HG--
extra : rebase_source : b3e3af3024622eadb6d061a640e93f6a83c9e34e
extra : topic : python-3.14