tests: adjust the timeout of test-bookmarks-corner-case.t
This is a standard "wait on file" situation, so we can use the standard tool for
that.

(We loose the reuse of the wait windows, but we gain standard tool usage that
automatically scale the timeout).

This should reduce the CI flakiness.

--HG--
branch : stable
extra : rebase_source : 54a9a046040a85dbc7905667b533ca636f3691fd
extra : topic : stabilize-test-bookmarks