file-index: represent empty tree as 0 bytes
This changes the file index to represent the empty tree as an empty file instead
of requiring a root node. This simplifies some code and complicates other code.
It has a few benefits which I think make it worthwhile:

* Nonexistent is treated the same as empty, like for the list and meta files.
* Adding the first path can be treated the same as others, no need to treat it
  like a vacuum in order to avoid a double root.
* It enables a follow-up change to make "base" non-optional in MutableTree.

The only user-facing change is that the docket for an empty file index shows a
0-byte tree file size, and debug::file-index --tree does not show a root node
for an empty file index.

--HG--
extra : rebase_source : aa8c573639b700143ddaafa6096d36b645fc9dc1
extra : topic : file-index.tokens