file-index: refactor size calculations
This refactors the calculations in the MutableTree serialize method in both
Python and Rust to use consistent terminology and hopefully be less confusing.
This is in preparation for a change to the format, which was easier to implement
after this improvement.

--HG--
extra : rebase_source : f79ebc08ee6cbbccb530c9e08a8783bf6cf31195
extra : topic : file-index.tokens