file-index: load root node eagerly
This makes the Python implementation of the file index load the root node
immediately after loading the tree file, instead of doing it at the start of
every lookup. This motivation is not performance but to align with how the Rust
implementation does it, to make it easier to change them together.

--HG--
extra : rebase_source : fcadf9f04d82f09502593af6b656eaa2fd34103c
extra : topic : file-index.tokens