file-index: clarify that path must be nonempty
Really there should never be an empty HgPath, but it is especially important for
FileIndex.add because there is nothing reasonable to do in that case and we
already assert on it in Rust. This just documents it better and adds more checks
to the Python implementation.

--HG--
extra : rebase_source : f307d90fe39cef0034d6b7bc446dd234e4cc8c61
extra : topic : file-index.tokens