pure-topo-branch: preset the heads revs for `pure_topo_branch`
In the pure-topological-case case we first retrieve the list of topological
heads as revision number and turn that into a list of node-id. So we can
directly fill the cache for `_head_revs` and `_open_head` to avoid having to
convert them back later.

Not having to turn them in node-id at all unless we actually needs it would be
nice, but we keep this changeset simple.

--HG--
extra : rebase_source : 18657c15b22018fb31732783291ca25258d20007
extra : topic : branchmap.v3-rev