branchmap: use a faster hashing for filtered and obsolete revision
This is a small change that should push the problem down the road enough that
the branch caching approach will have changed enough until this become a
noticeable problem again.

On my (beefy) developer machine, hashing all number from 0 to 5 000 000 see a
noticeable speedup with the current changeset. A quick and dirty check in Rust
is also included to highlight further potential speedup.

- sha1:  281ms
- crc32:  46ms
- rust:    1ms

CRC32 was mostly picked for availability and ease of use, but it also has the
advantage to allow for incremental update in case we want to start doing it.

--HG--
extra : rebase_source : 034d60634164fa8ce934d16070f464af0c5a0373
extra : topic : branchmap.hashing