tags: warm cache using fast delta when available
Rust knows how to compute a fast delta for revision close to each other, so we
leverage that to avoid restoring and parsing that many full text.

This is a low hanging fruit to improve the performance of this.

With this change, a `hg debugupgrade` run on a mozilla-try clone with  5793383
revision speedup from 10 minutes to 5 minutes.

(Note that this also means that if we were to expose the correct manifest diff
we compute for the rust update, we would have a fast and correct option to
compute the tags cache)

--HG--
extra : rebase_source : 82843eb1dfcb4d641b5bef4165a7ff2a860f3268
extra : topic : tag-rev-diff