changegroup: cleanup the old changegroup3 config
The format is no longer experimental and enabled by default for a long time. It
no longer seems relevant to keep that config around.

So we cleans it up to simplify the code and the tests.

--HG--
extra : rebase_source : ea10347d24bd8dd736da105a84598977bb21d781
extra : topic : 04-bundle