delta-search: optimize delta base if we won't reuse the compressed delta
We sometime skip the delta base optimization (by folding) to allow for a reuse
of the incoming compressed delta.

There is no use to refrain outself if we are not going to reuse the incoming
compression, so we stop doing so.

--HG--
extra : rebase_source : 40e96f945e4d75f56e3fb2461171a115e2921925
extra : topic : 04-bundle