file-index: always show root node in debug tree
This changes debug::file-index --tree to always show the root node. It was
skipped before because the root node pointer was not always valid. But now that
it is, there is no reason to hide it.

--HG--
extra : rebase_source : 900ae086fc80dcab1d520dcee11fae396de12df1
extra : topic : file-index.new-zero