file-index: use root token 0
This changes the file index to use 0 for the root node token instead of max u32.
This means tokens for real file paths start at 1 instead of 0.

The motivation for this is that if we ever have a use case for storing optional
tokens, it will be convenient to use 0 for the absent case, and not need to have
a special check for it when looking up its path.

Currently get_token("") and get_path(0) return None, so the root node is still
an implementation detail. In the future, we could consider making get_path(0)
return the empty path instead.

--HG--
extra : rebase_source : cd6c916c10cb3b839484b125de2b6237d2f32f89
extra : topic : file-index.new-zero