file-index: forbid singleton tree
This makes the file index reject a tree file that contains only a root node.
This ensures an empty tree is always be represented by an empty or absent file.
This was already the case before, but it was not checked.

The motivation: in some upcoming changes I inadvertently allowed the file to
write the singleton tree after removing paths, which caused a bug. It would be
possible to fix the bug by actually supporting the singleton tree. However, it
makes MutableTree additional/unused size logic more complicated, because the
root node might be leaf node. I think it's better to enforce a canonical
representation of the empty tree and then have simpler logic to deal with it.

--HG--
extra : rebase_source : 3a7690b09864612a47c53a988d565bb85cc5d6fc
extra : topic : file-index.new-zero