strip: avoid len(filelog) with file index
This changes strip to avoid checking len(filelog) == 0, since that can raise an
error with remotefilelog. I changed it to use the truthiness of the filelog
instead. For mercurial.filelog this is the same as len > 0. For remotefilelog,
it explicitly defines __bool__ to always be True.

I discovered this issue while running all the tests with the file index enabled
by default. It showed up in most of the test-remotefilelog-*.t tests.

--HG--
extra : rebase_source : a456d6d9ac121217f633e845dedec44ff939b850
extra : topic : file-index.bugs