file-index: change semantics of removing
This changes the file index remove method to do nothing if the given path
doesn't exist in the file index, rather than raising an error. This matches how
the fncache remove method behaves. If the caller really cares, they can assert
the path is in the file index before removing.

This fixes a bug when stripping revisions in a narrow repo. I added an example
in test-file-index.t which previously crashed. The problem is that repair.strip
processes all files affected by the revs to be stripped, and in a narrow repo
this can include paths excluded by the narrowspec, hence not in the file index.

This bug also occurred in more complicated uses of hg tracked by itself, since
it internally uses strip. I originally discovered the bug in test-narrow.t when
trying to run all tests with the file index enabled by default.

--HG--
extra : rebase_source : 13f872dbfbb886502432aaf27e241e5ac6def4eb
extra : topic : file-index.bugs