stream-clone: document the parallel-stream-bundle-processing
The feature have been around for a couple version and have been field tested for
multiple month. So, time to move it out of experimental and have it shows up in
the official documentation.

We fixes a couple of formatting errors and text inconsistency in the process.

--HG--
extra : topic : stream