contrib: stop installing Python 3.8 on Windows
We stopped supporting this a cycle or two ago.

If it's needed for some reason, the simplest way to get it is to run something
like this:

    winget install Python.Python.3.8 --custom "TargetDir=... InstallAllUsers=..."

... and include everything in the `Install-Python3` installer invocation after
the `/quiet` arg, and then use the installed python to run the downloaded `pip`
script.

--HG--
branch : stable
extra : rebase_source : d002047406ab028a4ed040bba9624ffbf5ad45a7
extra : topic : windows-deps-update