contrib: add some documentation about how to test Windows dependency updates
I needed a place to stash this info so I don't have to chase it down each time
an update is needed.  Also put a message at the end of the script when it
completes, because it swallows the output of the `hg clone` command, emits a few
empty lines, and then seems to hang forever when this is used to build a Docker
image.  (No, we shouldn't use this to build a Docker image, but it is very
convenient for repeatedly testing in a clean environment.)

--HG--
branch : stable
extra : rebase_source : 8a7a85ed3888c0034b8b2c96c25aee2fe6015ba0
extra : topic : windows-deps-update