shapes: don't panic when invalid paths are passed in
We were not supposed to get there, but we did, so might as well return results
instead of still panicking and gating more logic so the panic doesn't happen.
This is not a particularly hot path.

--HG--
branch : stable
extra : rebase_source : 92265455bad00efae54d213625ebce073c212299
extra : topic : shapes.oops