run-tests: include the `base_exec_prefix` directory in PATH on Windows
A recurring problem we have is that if python is not on PATH, the custom `hg.exe`
executable cannot find `python3.dll` and `python3X.dll`, and crashes early.  We
work around this in `setup.py` for `make local` by copying the files into the
root of the repo.  But that doesn't help `run-tests.py`, which builds and runs
the executable in an anonymous temp directory.  I've been getting away with it
for now by having Python 3.9 on PATH, but that means I can't use `py.exe` to
run tests with any other installed version.

Long term, we want to get rid of the custom executable and use the standard
python executable stub.  That would fix the issues with running a `pipx`
installed wheel.  But for now, sticking the root interpreter directory at the
front of PATH is an easy workaround to improve test coverage.

--HG--
branch : stable
extra : rebase_source : 8b2f8ee323ab7a04282f1456192d39bc1cdece3c
extra : topic : windows-tests-dll-load