test-censor: speed up file generation on Windows
In the new CI VM running on my laptop, this test finished in 1m30s.  On the
system hosting the VM, it took 368m.  Running with `--debug` showed lines like

        + for x in '`"$PYTHON" $TESTDIR/seq.py 0 50000`'
        + echo 'Password: hunter49976'

put out at the same time, a half second or so pause, and then a burst of the
same two lines that were next in sequence.  My best guess is that the antivirus
is scanning the file at least each time it is closed.  So instead, use inline
python and write the whole file in one shot.  Note that even though the new code
uses 50001 instead of 50000, the file is identical because `seq.py` adds 1 to
the second arg.  Now the host system can finish it in 1m55s.

--HG--
branch : stable
extra : rebase_source : 9215636849a20135ee9f50fdbfe14454019d3993
extra : topic : windows-test-fixes