procutil: avoid referencing `SIGKILL`, which is undefined on Windows
The original fix for this issue in 3f70ea5bcaeb was basically reverted in
4d924379c832, so this is more or less restoring what was.  The big difference is
this wasn't being called from core code before- only in tests.  We aren't
shipping `killdaemons.py` on Windows, so this will not work if it is ever called
outside of tests.

The signal argument it ditched, because `SIGKILL` can't be referenced on
Windows, and we aren't passing any other kill signal at the moment.

The last test to reference `ps_util` needs to be adjusted with this change
because changing it to call `procutil` first will break the test on Windows.

--HG--
branch : stable
extra : rebase_source : 894cbd06c7f6d8c1b483bd01b181ac018476279e
extra : topic : windows-sigkill