tests: use stdin to feed the PID kill list to `killdaemons.py` on Windows
We need to teach `procutil` to use this instead of using `signal.SIGKILL` for
Windows support.  However, we can't import this module there, because it creates
a cycle.  Next, we will need to copy the code there, but for now, simplify
things so that we don't create a temp file in core that relies on environment
variables that are setup by the test runner.

--HG--
branch : stable
extra : rebase_source : d100f02ea657fb8b510739d717ae44cf2003deff
extra : topic : windows-sigkill