tests: teach `killdaemons.py` to accept the PID list on stdin
We're going to have to invoke this from a core module, and while the expectation
is that it won't work outside of tests, not trying to setup a temp file that
relies on the test environment beforehand will be cleaner.  See the next commit
for details.

--HG--
branch : stable
extra : rebase_source : f44cd5b4d0aa78db421744e2a5e5711d902bed76
extra : topic : windows-sigkill