narrow: remove a test that doesn't have its place in core
This test already confesses to testing a feature that it itself introduces
and that cannot be in narrowhg proper (at least not without significant work).
It was imported from the `narrowhg` extension at the very beginning, which
understandably came at a more experimental stage and time, but it wouldn't
pass the bar today for reasons mentionned above.

An additional motivation is that the transaction business breaks down when
you squint too hard so it's better to get rid of it.

--HG--
branch : stable
extra : rebase_source : e3abb4c9d27dcd024b952d4402125c60104cb83c
extra : topic : narrowdirstate.fixup