tests: use inline python to create extremely long path on Windows
In order to exceed the MAX_PATH limit of 260 (and not fiddle with UNC paths),
a registry setting must be toggled on for the system, and the executable needs
the `longPathAware` setting in its application manifest.  The MSYS binaries are
old enough that neither `mkdir.exe` nor `touch.exe` have the application
manifest, so they can never exceed the limit.

The python 3.6 executable does have `longPathAware` in its application manifest,
so if `HKLM\SYSTEM\CurrentControlSet\Control\FileSystem@LongPathsEnabled` is set
to 1, this will create the long path and the file inside of it.  Only the wide
variants of the Win32 API support longer paths[1], but shockingly `hg.exe` can
handle this.  (This strongly implies that the bytes we pass to various file
functions are getting decoded to a wide character string before the wide API
functions are called.)

[1] https://learn.microsoft.com/en-us/windows/win32/fileio/maximum-file-path-limitation?tabs=powershell#functions-without-max_path-restrictions

--HG--
branch : stable
extra : rebase_source : fc5d1d8903865ab29c01d35fcc2de9a033a86695
extra : topic : windows-ci-fixes