repo-cache: fix using --chg with the rust extension (issue10048)
The repo cache cache was failing to update the opener on the Rust inner revlog.
For good reasons. It is unclear if this lack of override is bad at the inner
revlog level (we still needs it at the revlog level to properly detect update
without the lock).

The test seems happy with this patch, so lets says the change is good.

--HG--
branch : stable
extra : rebase_source : 3ead7062bc362b08b5d5d974a2cfe1b81f1198d2
extra : topic : ci-fixes