index: add an explicit parents method
This is the second step on the road to reduce use of the tuple form for a
revision entry.

The debug code wants to access the raw parent informations including when
invalid. So we introduce a dedicated method for this usecases. Otherwise the
test building corrupted changes can't debug the on-disk value.

I am not sure if I feel this is a valid use-case, but I would rather not derail
too much from my current goal.

--HG--
extra : rebase_source : b01789304fbdf2c6c99340e0f4678429e1992ab8
extra : topic : index-methods.parents