index: add an explicit method to retrieve a delta base
One more step on the road to reduce use of the tuple form for a
revision entry.

We take this as an opportunity to move the interpretation of the raw value
inside the method itself. For example, non-general delta, the stored value is
the root of the delta-chain and the delta-base is always `rev - 1` for delta.

That logic no longer needs to be copied around and can now be centralized.

--HG--
extra : rebase_source : f2ec7edbe4822d6b7cce2abfb7a3242aec3764fe
extra : topic : index-methods.delta-base