rust-fileindex: improve error reporting to Python
Before this change, the added test would traceback with a `ValueError`, which
isn't great. I had to adjust the Python code to also return an error and not
display a traceback in this case.

I don't plan on doing it for other such cases in the Python implementation
because it's purely here for completeness and we don't actually intend for
anyone to use the Python version of the fileindex for anything than a
best-effort compatibility layer.

--HG--
extra : rebase_source : 7a706cb551b8f9303f66b57609802ec905e1f721
extra : topic : rust.errors.fileindex