rust-errors: normalize hg-pyo3 dirstate exceptions
This is now all handled in a standard manner.

This affects the return code of rare errors in that it is more precise than
Python or C. Changing the Python implementation would be easy, but the C
implementation would be harder because it would require importing the
`error.InputError` class from the C module. All in all, I think this different
is minor enough that it's not worth bothering.

--HG--
extra : rebase_source : a0cdd15d914477520d1ebf1bc1ece7b470504c7d
extra : topic : rust.errors.valuerror