stream-clone: fix bug with file index and split filelog
This fixes a bug where stream cloning a file index repo did not work when there
was a split filelog. This was because the code path relied on the
RevlogStoreEntry property `_details`, which is not set with the file index. I
changed it to instead check based on `files` which is always present.

--HG--
branch : stable
extra : rebase_source : 2934aec5936853597aa26c69a05e30670bc127b0
extra : topic : file-index-clone-split