narrowspec: test comparing differently formatted narrowspecs
This adds a test for a shared repo whose narrowspec is formatted differently
than the store's narrowspec but contains the same includes and excludes.
Currently the Rust code fails because it checks that the two narrowspecs are
byte-for-byte identical. In a follow-up change I will make it instead compare
for logical equality the way Python does.

The motivation for this is that f176fb47c2bc changed the way narrowspecs are
formatted and this broke compatibility between 7.1 and 7.2.

--HG--
branch : stable
extra : rebase_source : 5bbbd9a8ce36576450d3cbb2851a1c25b365cefc
extra : topic : relax-narrowspec