ci: use a unique tags for the wheel builder
Having "macos, protected" as tags for the runner and for the job properly ensure
that only these runner will build the wheel. However, since the runner as the
"macos" tag it will also attempt to run the test job (with the "macos" tag
only).

This is not the goal, so we move to more unique tags to dispatch the jobs.

--HG--
branch : stable
extra : topic : protect-wheels