strip: forcibly write a fncache file consistent with the journal
In the middle of the strip, a transaction journal is written on disk to let an
interrupted strip to be recovered. However the fncache was not updated
accordily leading to possible inconsistent after the `hg recover`.

Seen inline comment for more details.

In the first test change show that we no longer have ghost entries in the
fncache, leading to a warning going away

In the second test change, the invalid revlog should not be there result in a
new warning about being missing from the fncache (as it should not be there at
all). This warning will be cleaned up later in that series.

--HG--
branch : stable
extra : rebase_source : 9414ead2666a2732c127123cb824684a6b70f5bc
extra : topic : fix-fncache-verify